Implemented standalone cache_utils outside of datagetter class
