+++ src/meteor/cmip6_meteor_data_getter.py
@@ -5,6 +5,7 @@ Module to get CMIP6 data and convert to format that can be used for METEOR
 import hashlib
 import logging
 import os
+import warnings
 
 import gcsfs
 import numpy as np
@@ -1360,6 +1361,10 @@ class Cmip6MeteorDataGetter:  # pylint: disable=too-many-instance-attributes
         """
         Get the standardized cache file path for a pattern scaling model.
 
+        .. deprecated:: 
+            Use :func:`meteor.cache_utils.get_pattern_scaling_cache_path` instead.
+            This method will be removed in a future version.
+
         Parameters
         ----------
         model_name : str
@@ -1375,39 +1380,18 @@ class Cmip6MeteorDataGetter:  # pylint: disable=too-many-instance-attributes
         -------
         str
             Full path to the cache file
-
-        Examples
-        --------
-        >>> data_getter = Cmip6MeteorDataGetter(exps=["piControl"], flds=["tas"])
-        >>> cache_path = data_getter.get_pattern_scaling_cache_path("CESM2")
-        >>> print(cache_path)
-        /path/to/.cache/trained_pattern_scaling_models/cmip6-CESM2-aer_pattern_scaling.pkl
         """
-        if cache_dir is None:
-            # Use default cache location in repository root
-            current_dir = os.path.dirname(os.path.abspath(__file__))
-            repo_root = current_dir
-            while repo_root != os.path.dirname(repo_root):
-                if any(
-                    os.path.exists(os.path.join(repo_root, marker))
-                    for marker in ["setup.py", ".git", "README.md"]
-                ):
-                    break
-                repo_root = os.path.dirname(repo_root)
-            cache_dir = os.path.join(
-                repo_root, ".cache", "trained_pattern_scaling_models"
-            )
+        warnings.warn(
+            "Cmip6MeteorDataGetter.get_pattern_scaling_cache_path() is deprecated. "
+            "Use meteor.cache_utils.get_pattern_scaling_cache_path() instead.",
+            DeprecationWarning,
+            stacklevel=2,
+        )
+        from . import cache_utils
 
-        os.makedirs(cache_dir, exist_ok=True)
-        if variable:
-            return os.path.join(
-                cache_dir,
-                f"cmip6-{model_name}-{scenario}-{variable}_pattern_scaling.pkl",
-            )
-        else:
-            return os.path.join(
-                cache_dir, f"cmip6-{model_name}-{scenario}_pattern_scaling.pkl"
-            )
+        return cache_utils.get_pattern_scaling_cache_path(
+            model_name, cache_dir, scenario, variable
+        )
 
     def validate_pattern_scaling_cache(
         self, cache_file, model_name, scenario="aer", variable=None
@@ -1415,8 +1399,9 @@ class Cmip6MeteorDataGetter:  # pylint: disable=too-many-instance-attributes
         """
         Validate a cached pattern scaling model file.
 
-        Checks if the cached pickle file exists, can be loaded, and contains
-        the expected model name and variable fields.
+        .. deprecated:: 
+            Use :func:`meteor.cache_utils.validate_pattern_scaling_cache` instead.
+            This method will be removed in a future version.
 
         Parameters
         ----------
@@ -1432,96 +1417,27 @@ class Cmip6MeteorDataGetter:  # pylint: disable=too-many-instance-attributes
         Returns
         -------
         tuple
-            (is_valid, cached_model, info_dict) where:
-            - is_valid: bool indicating if cache is valid
-            - cached_model: loaded model object if valid, None otherwise
-            - info_dict: dict with 'message', 'expected_name', 'found_name',
-              'expected_fields', 'found_fields'
-
-        Examples
-        --------
-        >>> data_getter = Cmip6MeteorDataGetter(exps=["piControl"], flds=["tas"])
-        >>> cache_file = data_getter.get_pattern_scaling_cache_path("CESM2")
-        >>> is_valid, model, info = data_getter.validate_pattern_scaling_cache(
-        ...     cache_file, "CESM2"
-        ... )
-        >>> if is_valid:
-        ...     print(f"✅ {info['message']}")
+            (is_valid, cached_model, info_dict)
         """
-        import pickle  # nosec B403
+        warnings.warn(
+            "Cmip6MeteorDataGetter.validate_pattern_scaling_cache() is deprecated. "
+            "Use meteor.cache_utils.validate_pattern_scaling_cache() instead.",
+            DeprecationWarning,
+            stacklevel=2,
+        )
+        from . import cache_utils
 
-        expected_name = f"cmip6-{model_name}-{scenario}"
-        expected_vars = set(self.flds)
-
-        info = {
-            "expected_name": expected_name,
-            "expected_fields": expected_vars,
-            "found_name": None,
-            "found_fields": set(),
-            "message": "",
-        }
-
-        # Check if file exists
-        if not os.path.exists(cache_file):
-            info["message"] = f"Cache file not found: {cache_file}"
-            return False, None, info
-
-        # Try to load and validate
-        try:
-            with open(cache_file, "rb") as f:
-                cached_data = pickle.load(f)  # nosec B301
-
-            # The MeteorPatternScaling.save_model() saves a dict, not the object itself
-            # Check if we loaded a dict (new format) or object (old format)
-            if isinstance(cached_data, dict):
-                # New format: dictionary with model data
-                info["found_name"] = cached_data.get("name", "unknown")
-                if "patternflds" in cached_data:
-                    info["found_fields"] = set(cached_data["patternflds"].keys())
-            else:
-                # Old format: try to get attributes from object
-                info["found_name"] = getattr(cached_data, "name", "unknown")
-                if hasattr(cached_data, "flds"):
-                    info["found_fields"] = set(cached_data.flds.keys())
-                elif hasattr(cached_data, "patternflds"):
-                    info["found_fields"] = set(cached_data.patternflds.keys())
-
-            # Validate model name
-            if info["found_name"] != expected_name:
-                info["message"] = (
-                    f"Model name mismatch: expected '{expected_name}', "
-                    f"found '{info['found_name']}'"
-                )
-                return False, None, info
-
-            # Validate fields exist
-            if not info["found_fields"]:
-                info["message"] = "Cached model missing field information"
-                return False, None, info
-
-            # Validate all expected fields are present
-            if not expected_vars.issubset(info["found_fields"]):
-                missing = expected_vars - info["found_fields"]
-                info["message"] = (
-                    f"Missing required fields: {missing}. "
-                    f"Expected {expected_vars}, found {info['found_fields']}"
-                )
-                return False, None, info
-
-            # Cache is valid
-            info["message"] = (
-                f"Cache valid: model={info['found_name']}, "
-                f"fields={list(info['found_fields'])}"
-            )
-            return True, cached_data, info
-
-        except Exception as e:
-            info["message"] = f"Error reading cache: {e}"
-            return False, None, info
+        return cache_utils.validate_pattern_scaling_cache(
+            cache_file, model_name, expected_fields=self.flds, scenario=scenario
+        )
 
     def get_noise_model_cache_path(self, model_name, variable_name, cache_dir=None):
         """
         Get the standardized cache file path for a noise model.
+
+        .. deprecated:: 
+            Use :func:`meteor.cache_utils.get_noise_model_cache_path` instead.
+            This method will be removed in a future version.
 
         Parameters
         ----------
@@ -1536,20 +1452,18 @@ class Cmip6MeteorDataGetter:  # pylint: disable=too-many-instance-attributes
         -------
         str
             Full path to the cache file
-
-        Examples
-        --------
-        >>> data_getter = Cmip6MeteorDataGetter(exps=["piControl"], flds=["tas"])
-        >>> cache_path = data_getter.get_noise_model_cache_path("CESM2", "tas")
-        >>> print(cache_path)
-        /path/to/noise_cache/CESM2_tas_noise_model.pkl
         """
-        if cache_dir is None:
-            # Use noise_cache in current working directory (notebook convention)
-            cache_dir = os.path.join(os.getcwd(), "noise_cache")
+        warnings.warn(
+            "Cmip6MeteorDataGetter.get_noise_model_cache_path() is deprecated. "
+            "Use meteor.cache_utils.get_noise_model_cache_path() instead.",
+            DeprecationWarning,
+            stacklevel=2,
+        )
+        from . import cache_utils
 
-        os.makedirs(cache_dir, exist_ok=True)
-        return os.path.join(cache_dir, f"{model_name}_{variable_name}_noise_model.pkl")
+        return cache_utils.get_noise_model_cache_path(
+            model_name, variable_name, cache_dir
+        )
 
     def validate_noise_model_cache(
         self,
@@ -1561,9 +1475,9 @@ class Cmip6MeteorDataGetter:  # pylint: disable=too-many-instance-attributes
         """
         Validate a cached noise model file.
 
-        Checks if the cached pickle file exists, can be loaded, and contains
-        the expected configuration (n_modes, lag_order, variable_name) and
-        required attributes (pca, varx_results).
+        .. deprecated:: 
+            Use :func:`meteor.cache_utils.validate_noise_model_cache` instead.
+            This method will be removed in a future version.
 
         Parameters
         ----------
@@ -1579,95 +1493,16 @@ class Cmip6MeteorDataGetter:  # pylint: disable=too-many-instance-attributes
         Returns
         -------
         tuple
-            (is_valid, cached_model, info_dict) where:
-            - is_valid: bool indicating if cache is valid
-            - cached_model: loaded MeteorNoiseGenerator if valid, None otherwise
-            - info_dict: dict with 'message', 'expected', 'found' information
-
-        Examples
-        --------
-        >>> data_getter = Cmip6MeteorDataGetter(exps=["piControl"], flds=["tas"])
-        >>> cache_file = data_getter.get_noise_model_cache_path("CESM2", "tas")
-        >>> is_valid, model, info = data_getter.validate_noise_model_cache(
-        ...     cache_file, "tas", n_modes=40, lag_order=2
-        ... )
-        >>> if is_valid:
-        ...     print(f"✅ {info['message']}")
+            (is_valid, cached_model, info_dict)
         """
-        # Import here to avoid circular dependency
-        from meteor.noise_generator import MeteorNoiseGenerator
+        warnings.warn(
+            "Cmip6MeteorDataGetter.validate_noise_model_cache() is deprecated. "
+            "Use meteor.cache_utils.validate_noise_model_cache() instead.",
+            DeprecationWarning,
+            stacklevel=2,
+        )
+        from . import cache_utils
 
-        info = {
-            "expected": {
-                "variable_name": variable_name,
-                "n_modes": n_modes,
-                "lag_order": lag_order,
-            },
-            "found": {},
-            "message": "",
-        }
-
-        # Check if file exists
-        if not os.path.exists(cache_file):
-            info["message"] = f"Cache file not found: {cache_file}"
-            return False, None, info
-
-        # Try to load and validate
-        try:
-            noise_model = MeteorNoiseGenerator(n_modes=n_modes, lag_order=lag_order)
-            noise_model.load_model(cache_file)
-
-            # Extract found information
-            info["found"]["n_modes"] = getattr(noise_model, "n_modes", None)
-            info["found"]["lag_order"] = getattr(noise_model, "lag_order", None)
-            info["found"]["variable_name"] = getattr(noise_model, "variable_name", None)
-
-            # Validate n_modes
-            if not hasattr(noise_model, "n_modes") or noise_model.n_modes != n_modes:
-                info["message"] = (
-                    f"n_modes mismatch: expected {n_modes}, "
-                    f"found {info['found']['n_modes']}"
-                )
-                return False, None, info
-
-            # Validate lag_order
-            if (
-                not hasattr(noise_model, "lag_order")
-                or noise_model.lag_order != lag_order
-            ):
-                info["message"] = (
-                    f"lag_order mismatch: expected {lag_order}, "
-                    f"found {info['found']['lag_order']}"
-                )
-                return False, None, info
-
-            # Validate variable_name
-            if (
-                not hasattr(noise_model, "variable_name")
-                or noise_model.variable_name != variable_name
-            ):
-                info["message"] = (
-                    f"variable_name mismatch: expected '{variable_name}', "
-                    f"found '{info['found']['variable_name']}'"
-                )
-                return False, None, info
-
-            # Validate required attributes
-            required_attrs = ["pca", "varx_results"]
-            missing_attrs = [
-                attr for attr in required_attrs if not hasattr(noise_model, attr)
-            ]
-            if missing_attrs:
-                info["message"] = f"Missing required attributes: {missing_attrs}"
-                return False, None, info
-
-            # Cache is valid
-            info["message"] = (
-                f"Cache valid: variable={variable_name}, "
-                f"n_modes={n_modes}, lag_order={lag_order}"
-            )
-            return True, noise_model, info
-
-        except Exception as e:
-            info["message"] = f"Error loading cache: {e}"
-            return False, None, info
+        return cache_utils.validate_noise_model_cache(
+            cache_file, variable_name, n_modes, lag_order
+        )
